python-ai/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
httpx==0.26.0
feedparser==6.0.11
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyyaml==6.0.1
